+++ drivers/socketcan/socketcan.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016 UAVCAN Team
+ * Copyright (c) 2016-2018 UAVCAN Team
  *
  * Distributed under the MIT License, available in the file LICENSE.
  *
@@ -22,28 +22,33 @@ typedef struct
 
 /**
  * Initializes the SocketCAN instance.
+ * Returns 0 on success, negative on error.
  */
-int socketcanInit(SocketCANInstance* out_ins, const char* can_iface_name);
+int16_t socketcanInit(SocketCANInstance* out_ins, const char* can_iface_name);
 
 /**
  * Deinitializes the SocketCAN instance.
+ * Returns 0 on success, negative on error.
  */
-int socketcanClose(SocketCANInstance* ins);
+int16_t socketcanClose(SocketCANInstance* ins);
 
 /**
  * Transmits a CanardCANFrame to the CAN socket.
  * Use negative timeout to block infinitely.
+ * Returns 1 on successful transmission, 0 on timeout, negative on error.
  */
-int socketcanTransmit(SocketCANInstance* ins, const CanardCANFrame* frame, int timeout_msec);
+int16_t socketcanTransmit(SocketCANInstance* ins, const CanardCANFrame* frame, int32_t timeout_msec);
 
 /**
  * Receives a CanardCANFrame from the CAN socket.
  * Use negative timeout to block infinitely.
+ * Returns 1 on successful reception, 0 on timeout, negative on error.
  */
-int socketcanReceive(SocketCANInstance* ins, CanardCANFrame* out_frame, int timeout_msec);
+int16_t socketcanReceive(SocketCANInstance* ins, CanardCANFrame* out_frame, int32_t timeout_msec);
 
 /**
  * Returns the file descriptor of the CAN socket.
+ * Can be used for external IO multiplexing.
  */
 int socketcanGetSocketFileDescriptor(const SocketCANInstance* ins);
